% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearings.R
\name{calc_bearings}
\alias{calc_bearings}
\title{bearing calculation}
\usage{
calc_bearings(
  projList,
  animal,
  path_to_data,
  antennas,
  dbLoss,
  calibrate,
  uav = FALSE,
  version
)
}
\arguments{
\item{animal}{list, generatet by initanimal}

\item{path_to_data}{string, path to filtered file}

\item{antennas}{file, data.frame with antenna information (Lat,Lon,orientation)}

\item{dbLoss}{num, gain loss from 0 to 90° antenna orientation}

\item{calibrate}{logical, if TRUE correction values are applied}

\item{uav}{logical, if TRUE antenna orientation based on uav pictures is used}

\item{version}{string, string pattern specifying the version of data processing}
}
\description{
calculates bearings using linear approximation and acos
}
\author{
Jannis Gottwald
}
