% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision_detector.R
\name{collision_detector}
\alias{collision_detector}
\title{identify signal collision}
\usage{
collision_detector(animal10, animal20, animal40, td10, td20, version)
}
\arguments{
\item{animal10}{list, generatet by initanimal-meta information of tag with 10ms duration}

\item{animal20}{list, generatet by initanimal-meta information of tag with 20ms duration}

\item{animal40}{list, generatet by initanimal-meta information of tag with 40ms duration}

\item{td10}{num, expected timedifference between signals of tag with 10ms duration}

\item{td20}{num, expected timedifference between signals of tag with 20ms duration}

\item{version}{string,string pattern specifying the version of data processing}
}
\description{
find collsions between tags of same frequency but different signal duration by finding gaps in the timeseries
}
\author{
Jannis Gottwald
}
