% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_interactive.R
\name{filter_interactive}
\alias{filter_interactive}
\title{filter data interactive}
\usage{
filter_interactive(projList, animal, path_to_data = ".", nth)
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{animal}{list, generated by initanimal function}

\item{nth}{num, subset data for plotting using only every nth column}

\item{station}{string, path to awk filtered file}
}
\description{
Interactive filtering of awk filtered data. Awk files may miss colnames. make shure they have colnames(data)="timestamp", "samples", "duration", "signal_freq", "signal_bw", "max_signal","noise", "receiver", "station"
}
\author{
Jannis Gottwald
}
