% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts-Parent_Mild_0.R
\name{get_parent}
\alias{get_parent}
\title{Returns parent survey data}
\usage{
get_parent(d, schedule = "baseline", severity = "mild/moderate")
}
\arguments{
\item{d}{The full data. Output from \link{get_rts_data}}

\item{schedule}{The assessment schedule. Defaults to \code{baseline}.
Should be one of \code{"baseline"}, \code{"2"}, \code{"4"}, \code{"8"},
\code{"12"}, \code{"24"}, or \code{"end"}.}

\item{severity}{The severity of the TBI. Defaults to \code{"mild/moderate"}.
Should be one of \code{"mild/moderate"} or \code{"severe"}.}
}
\description{
Returns parent survey data
}
