% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rts_data.R
\name{get_rts_data}
\alias{get_rts_data}
\title{Return full REDCap data}
\usage{
get_rts_data(redcap_api_token)
}
\arguments{
\item{redcap_api_token}{The API token from REDCap, passed as a string}
}
\value{
A data frame with the raw data
}
\description{
Returns the full data across all surveys
}
