% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{get_survey}
\alias{get_survey}
\title{Returns parent survey data}
\usage{
get_survey(d, severity = "mild", schedule = "baseline", survey = "parent")
}
\arguments{
\item{d}{The full data. Output from \code{\link[=get_rts_data]{get_rts_data()}}}

\item{severity}{The severity of the TBI. Defaults to \code{"mild"}.
Should be one of \code{"mild"} or \code{"moderate/severe"}.}

\item{schedule}{The assessment schedule. Defaults to \code{baseline}.
Should be one of \code{"baseline"}, \code{"2"}, \code{"4"}, \code{"8"},
\code{"12"}, \code{"24"}, \code{"36"}, \code{"52"}, \code{"104"} or
\code{"end"}.}

\item{survey}{The survey to retrieve. Should be one of \code{"parent"},
\code{"student"}, \code{"teacher1"}, or \code{"teacher2"}.}
}
\description{
Returns parent survey data
}
