% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampel_time.R
\name{hampel}
\alias{hampel}
\title{hampel filter}
\usage{
hampel(path_to_data, col, k, t0 = 0.5, animal, ncores)
}
\arguments{
\item{path_to_data}{string, dpath to file containing bearings}

\item{col}{string, string indicating the column containing bearings}

\item{k}{numeric, window size n seconds}

\item{t0}{numeric,threshold for median filter. A high threshold makes the filter more forgiving, a low one will declare more points to be outliers. Default=0.5}
}
\description{
detect outliers in bearings using hampel filter in a moving time window
}
\author{
Jannis Gottwald
}
