% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initAnimal.R
\name{initAnimal}
\alias{initAnimal}
\title{Init Animal Tracking}
\usage{
initAnimal(projList, animalID, freq, start, end)
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{animalID}{string, label of the tagged individual}

\item{freq}{num, tag frequency (khz)}

\item{start}{string, start of tracking YYYY-MM-DD}

\item{end}{string, end of tracking YYYY-MM-DD}
}
\description{
Creates directories and a R list with all the needed tRackIT information for one animal
}
\author{
Marvin Ludwig & Jannis Gottwald
}
