% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProject.R
\name{initProject}
\alias{initProject}
\title{Init radio-tracking project}
\usage{
initProject(projroot = ".", logger_data_raw)
}
\arguments{
\item{projroot}{string, project directory}

\item{logger_data_raw}{string, path to the full RTS dataset}
}
\description{
Creates directories and a R list with all the needed tRackIT information for one animal
}
\author{
Marvin Ludwig & Jannis Gottwald
}
