% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_vars.R
\name{machine_vars}
\alias{machine_vars}
\title{create variables for ML}
\usage{
machine_vars(path_to_data, animal, w = c(5, 11))
}
\arguments{
\item{path_to_data}{string, path to station data}

\item{animal}{list, generated by initAnimal function}

\item{w}{numeric, window size for rollaply}
}
\description{
Creates different variables from raw VHF signals to support machine learning
}
\author{
Jannis Gottwald
}
