% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gps.R
\name{match_gps}
\alias{match_gps}
\title{filter data interactivematch to gps data}
\usage{
match_gps(data, antennas, gps, td_to_gps)
}
\arguments{
\item{data}{data.frame, data.frame with bearings}

\item{gps}{data.frame, gps data}

\item{td_to_gps}{numeric, timedifference between bearing timestamps (UTC) and gps timestamp}

\item{nth}{num, subset data for plotting using only every nth column}
}
\description{
match bearings to gps for accuracy assessment
}
\author{
Jannis Gottwald
}
