% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bearings.R
\name{predict_bearings}
\alias{predict_bearings}
\title{bearing prediction}
\usage{
predict_bearings(data, animal, mxNA, path_to_data)
}
\arguments{
\item{animal}{list, generatet by initanimal}

\item{mxNA}{num, may number of missing signals accepted}

\item{path_to_data}{string, path to filtered file}
}
\description{
predict bearings using circtree model
}
\author{
Jannis Gottwald
}
