% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logger_data.R
\name{read.logger.data}
\alias{read.logger.data}
\title{create file and parameter catalogue from raw logger data}
\usage{
read.logger.data(projList, station = ".", collection = ".")
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{station}{string, station of hich the files should be catagolized}

\item{collection}{string, path to data collection in station folder from which the files should be catagolized}
}
\description{
Creates a catalogue of all raw logger files in folder
}
\author{
Jannis Gottwald
}
