% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts-package.R
\docType{package}
\name{rts-package}
\alias{rts}
\alias{rts-package}
\title{rts: Returns data from the Return to School project}
\description{
Returns data from REDCap API for the Return to School project at 
             the University of Oregon on traumatic brain injury. Also includes 
             a few helper functions for working with the data.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

Authors:
\itemize{
  \item Jeff Gau \email{jeffg@uoregon.edu}
}

}
\keyword{internal}
