% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match_station.R
\name{time_match_station}
\alias{time_match_station}
\title{time match logger data}
\usage{
time_match_station(animal, method, collision = FALSE)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{method}{string, which bearing calculation method ("linear", "acos", "ML")}

\item{collision}{logical, if collision corrected data is used a suffix is added to resulting file name}
}
\description{
timematch of simoultaneously received signals by stations
}
\author{
Jannis Gottwald
}
