% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_rolling_fun.R
\name{bearing_rolling_fun}
\alias{bearing_rolling_fun}
\title{triangulation}
\usage{
bearing_rolling_fun(
  animal,
  antennas,
  version = "",
  tw = 5,
  epsg = 4326,
  maxDB = 60,
  maxNA = 2,
  GPS = FALSE,
  plot_results = FALSE,
  x = "ML",
  xlim = c(8.665209, 8.700571),
  ylim = c(50.826948, 50.849172),
  keep_cores = 2
)
}
\arguments{
\item{animal}{list, generatet by initanimal}

\item{antennas}{data.frame, Data.frame with antenna information}

\item{version}{string, character string for identification}

\item{tw}{num, time window in seconds: Locations are calculated using all bearings tw sec before and after timestamp in row}

\item{epsg}{num, epsg code for coordinate refernce system in use (must be metric). Lat/Lon is default}

\item{maxDB}{num, minimum value for received maximum signal strength. default=60}

\item{GPS}{logical, if true reference gps data is expected in cols lon/lat}

\item{plot_results}{logical, wether to plot bearings during processing or not. Slows down the calculation substantially}

\item{x}{string, name of column with bearings. x="bearings_filtered" if hampel filtered bearings are used}

\item{xlim}{numeric, x limits for plot}

\item{ylim}{numeric, y limits for plot}

\item{method}{string, pattern indicating whether ML bearings or bearings based on physical model should be selected}

\item{maxNa}{num, max for missing receiver per row. default=2}
}
\description{
triangulate using telemtr package (not on CRAN) by barry rowlingson (https://github.com/barryrowlingson/telemetr). Methods are described here: https://www.jstor.org/stable/1268030?seq=1#metadata_info_tab_contents. Use remote::install_github("barryrowlingson/telemetr") and set "no errors from warnings" to TRUE before installing the package - Sys.setenv("R_REMOTES_NO_ERRORS_FROM_WARNINGS"=TRUE).
}
\author{
Jannis Gottwald
}
